// src/exchanges/bitfinex/enums/BitfinexOrderTypeEnum.ts
export enum BitfinexOrderTypeEnum {
  MARKET = 'MARKET',
  LIMIT = 'LIMIT',
  STOP = 'STOP',
  STOP_LIMIT = 'STOP LIMIT',
  TRAILING_STOP = 'TRAILING STOP',
  FOK = 'FOK',
  IOC = 'IOC',
  EXCHANGE_MARKET = 'EXCHANGE MARKET',
  EXCHANGE_LIMIT = 'EXCHANGE LIMIT',
  EXCHANGE_STOP = 'EXCHANGE STOP',
  EXCHANGE_STOP_LIMIT = 'EXCHANGE STOP LIMIT',
  EXCHANGE_TRAILING_STOP = 'EXCHANGE TRAILING STOP',
  EXCHANGE_FOK = 'EXCHANGE FOK',
  EXCHANGE_IOC = 'EXCHANGE IOC',
}